Use i32 as much as possible (#393)
Seems like the tiniest of slow downs. Can't really figure out why. On the whole, I think I'm using _less_ memory by turning a lot of the `usize` into `i32`. Maybe some unfortunate cache-related side effects?

```
Elo   | -2.53 +- 2.01 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | -2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 41942 W: 11910 L: 12216 D: 17816
Penta | [739, 4876, 10011, 4642, 703]
https://chess.samroelants.com/test/1226/
```
https://chess.samroelants.com/test/1226/

bench 3241984

Reviewed-on: https://git.samroelants.com/sroelants/simbelmyne/pulls/393

+++ crates/cli/src/cli/bench.rs
@@ -72,7 +72,7 @@ pub struct BenchHarness {
 }
 
 impl BenchHarness {
-  pub fn new(depth: usize, nodes: Option<usize>) -> Self {
+  pub fn new(depth: i32, nodes: Option<usize>) -> Self {
     let tc = if let Some(nodes) = nodes {
       TimeControl::Nodes(nodes)
     } else {
@@ -103,7 +103,7 @@ impl BenchHarness {
   }
 }
 
-pub fn run_bench(depth: usize, nodes: Option<usize>) {
+pub fn run_bench(depth: i32, nodes: Option<usize>) {
   let mut harness = BenchHarness::new(depth, nodes);
 
   for fen in POSITIONS {
